+++ src/helper-modules/bpmn-helper/src/getters.js
@@ -783,7 +783,7 @@ async function getMilestonesFromElementById(bpmn, elementId) {
  * An object containing information about a variable that might exist during the instance of a process
  *
  * @typedef Variable
- * 
+ *
  * @type {object}
  * @property {string} name - variable name
  * @property {string} [description] - a description of the variable
@@ -804,10 +804,28 @@ function getVariablesFromElement(element) {
   let variables = [];
   if (element.extensionElements && Array.isArray(element.extensionElements.values)) {
     const variablesElement = element.extensionElements.values.find(
-      (child) => child.$type == 'proceed:Variables'
+      (child) => child.$type == 'proceed:Variables',
     );
     if (variablesElement && variablesElement.variable) {
-      variables = variablesElement.variable;
+      variables = variablesElement.variable.map(
+        ({
+          name,
+          description,
+          dataType,
+          defaultValue,
+          requiredAtInstanceStartup,
+          enum: e,
+          const: c,
+        }) => ({
+          name,
+          description,
+          dataType,
+          defaultValue,
+          requiredAtInstanceStartup,
+          enum: e,
+          const: c,
+        }),
+      );
     }
   }
